refactor: resolve conflicts
