+++ .env.example
@@ -10,14 +10,13 @@ LINE_KEY
 LINE_SECRET
 GOOGLE_KEY
 GOOGLE_SECRET
-
+MAILGUN_API_URL
+MAILGUN_API_KEY
+MAILGUN_SENDER_DOMAIN
+EMAIL_HOST_USER
+EMAIL_FROM
+DEFAULT_FROM_EMAIL
 AWS_ACCESS_KEY_ID
 AWS_SECRET_ACCESS_KEY
 AWS_STORAGE_BUCKET_NAME
 AWS_S3_REGION_NAME
-
-
-MAILGUN_API_URL
-MAILGUN_API_KEY
-EMAIL_FROM
-
