+++ apps/companies/admin.py
@@ -1,3 +1,1 @@
 from django.contrib import admin
-
-# Register your models here.
