+++ apps/posts/models.py
@@ -1,4 +1,3 @@
-from django.core.validators import MaxValueValidator, MinValueValidator
 from django.db import models
 
 from apps.companies.models import Company
