+++ apps/users/templates/users/password_reset_done.html
@@ -3,7 +3,7 @@
   <section class="w-full max-w-md p-6 bg-white rounded-lg shadow-lg">
     <h1 class="text-2xl font-bold text-center">密碼重置郵件已發送</h1>
     <p class="mt-4 text-center">
-      如果您的電子郵件地址存在，您將會收到一封包含重置鏈接的郵件。
+      如果您的電子郵件地址存在，您將會收到一封包含重置密碼的郵件。
     </p>
     <div class="mt-4 text-center">
       <a href="{% url 'users:sign_in' %}" class="btn btn-primary">返回登入</a>
