+++ apps/users/urls.py
@@ -1,5 +1,4 @@
 from django.urls import path
-from django.views.generic import TemplateView
 
 from . import views
 from .views import PasswordResetDoneView, PasswordResetView
