+++ config/urls.py
@@ -19,6 +19,8 @@ from debug_toolbar.toolbar import debug_toolbar_urls
 from django.contrib import admin
 from django.urls import include, path
 
+from lib.utils.env import is_dev
+
 urlpatterns = [
     path("companies/", include("apps.companies.urls")),
     path("admin/", admin.site.urls),
@@ -27,4 +29,7 @@ urlpatterns = [
     path("posts/", include("apps.posts.urls")),
     path("social-auth/", include("social_django.urls", namespace="social")),
     path("resumes/", include("apps.resumes.urls")),
-] + debug_toolbar_urls()
+]
+
+if is_dev():
+    urlpatterns += debug_toolbar_urls()
