+++ templates/shared/navbar.html
@@ -1,13 +1,13 @@
 <nav class="flex items-center justify-between p-4 shadow navbar bg-base-100">
-  <!-- 左上角：回首頁 -->
+
   <div class="flex items-center">
     <a href="/" class="text-blue-500 hover:underline">回到首頁</a>
   </div>
-  <!-- 中間正上方：WorkNet -->
+
   <div class="flex-1 text-center">
     <a href="{% url 'users:index' %}" class="text-xl font-bold">WorkNet</a>
   </div>
-  <!-- 右上角：登入/註冊 或 登出 -->
+
   <div class="flex items-center gap-4">
     {% if request.user.is_authenticated %}
     <div class="flex gap-2 text-sm">
